compFitDatabase.sql
This is the first draft of the mysql database creation script. Need to do more research on storing images in a database (for the avatar), what the salt for the passwords would look like, and we need to flesh out team_progress more. Also on the name of the database, whether we want capitalization or not, and what exactly we want to call the script file. No mock data yet.