Endpoints - POST user, some team
I added a post for user using the userPost.html. Probably need to
change this based on what GUI does. I also added a delete to
user/user_id that can be carried out with a command on console (curl –i
–X DELETE http://zero-to-slim.dev/user/{user_id}). Works for me. I
don't know what to return for this. Added some get methods for team.